rename writable modules to writeable
